move server files to src/plugins
