+++ src/plugins/newsfeed/server/plugin.ts
@@ -1,0 +1,45 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { Plugin, PluginInitializerContext, CoreSetup, Logger } from '../../../../src/core/server';
+import { defineRoutes } from './routes';
+
+export type Setup = void;
+export type Start = void;
+
+export class NewsfeedServerPlugin implements Plugin<Setup, Start> {
+  private readonly logger: Logger;
+  private readonly kibanaVersion: string;
+
+  constructor(initializerContext: PluginInitializerContext) {
+    this.logger = initializerContext.logger.get();
+    this.kibanaVersion = initializerContext.env.packageInfo.version;
+  }
+
+  setup(core: CoreSetup) {
+    defineRoutes({
+      router: core.http.createRouter(),
+      kibanaVersion: this.kibanaVersion,
+    });
+  }
+
+  start(core: CoreSetup) {
+    this.logger.debug('Starting plugin');
+  }
+}
